Extract out inline class.
